updated readme and some scripts (cleanup and container forking)

+++ b2d/bin/cleancontainers.sh
@@ -4,7 +4,15 @@ counter=1
 numberhosts=7
 hostname=dkrctrhost
 ctrname=dkrctrnm
+numberctrs=`docker ps -a | grep $ctrname | tr -s " " | cut -d " " -f 1 | wc -l`
 ctrlist=`docker ps -a | grep $ctrname | tr -s " " | cut -d " " -f 1`
-docker rm $ctrlist
 
+if [[ $numberctrs -gt 0 ]] ; then 
+	echo "Stopping containers ..."
+	docker stop $ctrlist
+	echo "Removing containers ..."
+	docker rm $ctrlist
+else
+	echo "No container to stop / remove !"
+fi
 exit 0
